Start testing on PHP 7.1. Fix version of PHPUnit in use.
PHP 7.1 has been available for a while now and all of my testing
indicates it works fine.

The version of PHPUnit has been fixed at the latest v4 edition. This
allows us to continue supporting PHP 5.4, and also lets us use the
backported PHPUnit 6 classnames for test cases, to make the eventual
upgrade easier. If we were to drop PHP 5.4 and 5.5 support, we could
move to PHPUnit v5.

The composer.lock file is now committed, to ensure everyone is using the
same test framework. This does not pose any issues for users of this
library, as composer ignores composer.lock files and requires-dev in
dependencies.