Fix deprecation warnings in PHP 8.1
According to the PHP 8.1 changelog:

> Passing null to non-nullable internal function parameters is
> deprecated.

There are a handful of places in BBCode.php that pass null to the
`strlen` function, which results in the following warnings:

> Deprecated: strlen(): Passing null to parameter #1 ($string) of
> type string is deprecated

This affects 3 functions:

1) cleanupWSByPoppingStack
2) cleanupWSByEatingInput
3) cleanupWSByIteratingPointer

In the cases of 1) and 2), passing `$pattern = null` is a no-op, and
therefore we can safely omit these calls entirely. This is because we
immediately trigger the condition:

    if (strlen($pattern) <= 0)
        return;

In the case of 3), passing `$pattern = null` returns the 2nd argument
passed to the function, so we can simply use this value directly
instead. This is because we immediately trigger the condition:

    if (strlen($pattern) <= 0)
        return $pos;

These changes should fix the deprecation warnings, because we will no
longer pass null to `strlen`, while also leaving the logic unchanged.