Fix [list] element in PHP 8.0
The BBCodeLibrary::doList function was affected by a change to how
string to number comparison works in PHP 8.0.

In PHP 7.4 and earlier:

    0 == "foo"  // evalulates to true

In PHP 8.0+:

    0 == "foo"  // evaluates to false

See the RFC at https://wiki.php.net/rfc/string_to_number_comparison.

We fix this by changing:

    strlen($default) == ""

to:

    strlen($default) == 0

as the output of `strlen` is always an int and should not be compared
to a string.

The PHP 8.0 test is re-enabled in Circle CI.